Autopackage for Debian
This commit allows for the autopackaging of tumble into a debian package.
Simply typing `make deb` will do the needful. The package is versioned using a
`git describe` call which does require annoated tags to base the description
upon.

This should replace the existing deployment mechanism for tumble, pointing at
tumble.wcyd.org repository, as that is now behind this repository. Since that
repo contains many hard-coded assumptions, this repo is to be preferred.