Render twitter links as actual twitter links.
Prior to this small patch set, twitter links were just normal string uris on
the tumble page. It was boring. I wanted them to be like normal embedded tweets
with the ability to clink on them, respond to them, follow people, etc. This is
what is happening now.

To create this, I wrote a small external utility in golang (because I wanted
to). It simply takes a twitter uri and converts it into the HTML required to
render the tweet appropriately. The Javascript fun had to be added to the main
tumble render template as well.